Moved and deleted tests
