+++ dockstore-cli-integration-testing/src/test/java/io/dockstore/client/cli/BasicIT.java
@@ -718,20 +718,6 @@ public class BasicIT extends BaseIT {
 
 
     /**
-     * Will test attempting to manually publish a Dockerhub/Bitbucket entry using incorrect CWL and/or dockerfile locations
-     */
-    @Ignore
-    public void testDockerhubBitbucketWrongStructure() {
-        // Todo : Manual publish entry with wrong cwl and dockerfile locations, should not be able to manual publish
-        systemExit.expectSystemExitWithStatus(Client.GENERIC_ERROR);
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "manual_publish", "--registry",
-            Registry.DOCKER_HUB.name(), Registry.DOCKER_HUB.toString(), "--namespace", "dockstoretestuser", "--name",
-            "dockerhubandbitbucketalternate", "--git-url", "git@bitbucket.org:DockstoreTestUser/quayandbitbucketalterante.git",
-            "--git-reference", "master", "--toolname", "alternate", "--cwl-path", "/Dockstore.cwl", "--dockerfile-path", "/Dockerfile",
-            "--script" });
-    }
-
-    /**
      * Checks that you can manually publish and unpublish a Dockerhub/Bitbucket duplicate if different toolnames are set (but same Path)
      */
     @Test
